Made EPF link downloadable
Alt-clicking the EPF file link will download a file with HTML code that will not work when imported into Eclipse (because it takes the user to a github page, not to the raw code). I have replaced the link with one that shows the raw code and I have added a way to download it directly. This allows to alt-click as well.